webpage-ai-parser/requirements.txt
fastapi
uvicorn
playwright
beautifulsoup4
groq
python-dotenv
requests
